requirements.txt
psycopg2-binary
python-dotenv
flask-restx
flask-jwt-extended
flask